CF-46 encapsulated login UI and implemented logout button
