+++ src/components/UI/PageWrapper/PageWrapperStyles.js
@@ -1,16 +1,13 @@
 import { styled } from '@mui/system';
 
-
 const WrapperDiv = styled(`div`)({
   height: '100%',
   width: '91%',
   position: 'absolute',
-  left: '180px',
+  left: '150px',
   top: '5rem',
   backgroundColor: 'none',
   paddingTop: '100px',
 });
-
-
 
 export { WrapperDiv };
